requirements.txt
Django==1.6.5
Markdown==2.4.1
argparse==1.2.1
boto==2.29.1
django-filter==0.7
django-oauth2-provider==0.2.6.1
djangorestframework==2.3.14
shortuuid==0.4.2
wsgiref==0.1.2
